# openapi.json
{
...
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/v1/namespaces": {
      "post": {
        "operationId": "createNamespace",
        "summary": "",
        "description": "Create namespace\nRole required: NS_MANAGER",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNamespaceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The namespace has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Namespace"
                }
              }
            }
          }
        },
        "tags": ["namespace"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "listNamespaces",
        "summary": "",
        "description": "List namespaces\nRole required: NS_MANAGER",
        "parameters": [
          {
            "name": "parent",
            "required": false,
            "in": "query",
            "description": "parent",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "parent_scope",
            "required": false,
            "in": "query",
            "description": "parent scope",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "_sort",
            "required": false,
            "in": "query",
            "description": "排序参数",
            "schema": {
              "enum": ["createAt", "-createAt", "updateAt", "-updateAt"],
              "type": "string"
            }
          },
          {
            "name": "name_like",
            "required": false,
            "in": "query",
            "description": "名称 模糊查询",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "labels",
            "required": false,
            "in": "query",
            "description": "标签",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "_limit",
            "required": false,
            "in": "query",
            "description": "分页大小",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "_offset",
            "required": false,
            "in": "query",
            "description": "分页偏移",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A paged array of namespaces.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Namespace"
                  }
                }
              }
            }
          }
        },
        "tags": ["namespace"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/auth/v1/namespaces/{namespaceIdOrNs}": {
      "get": {
        "operationId": "getNamespace",
        "summary": "",
        "description": "Find namespace by\nRole required: NS_MANAGER",
        "parameters": [
          {
            "name": "namespaceIdOrNs",
            "required": true,
            "in": "path",
            "description": "Namespace id or ns, should encodeURIComponent",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The namespace with expected id or ns.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Namespace"
                }
              }
            }
          }
        },
        "tags": ["namespace"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/auth/v1/namespaces/{namespaceId}": {
      "patch": {
        "operationId": "updateNamespace",
        "summary": "",
        "description": "Update namespace\nRole required: NS_MANAGER",
        "parameters": [
          {
            "name": "namespaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNamespaceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The namespace updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Namespace"
                }
              }
            }
          }
        },
        "tags": ["namespace"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "deleteNamespace",
        "summary": "",
        "description": "Delete namespace\nRole required: NS_MANAGER",
        "parameters": [
          {
            "name": "namespaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No content."
          }
        },
        "tags": ["namespace"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/auth/v1/scopes": {
      "get": {
        "operationId": "listScopes",
        "summary": "",
        "description": "List namespaces scopes\nRole required: AUTH_MANAGER",
        "parameters": [
          {
            "name": "parent",
            "required": false,
            "in": "query",
            "description": "parent",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "_sort",
            "required": false,
            "in": "query",
            "description": "排序参数",
            "schema": {
              "enum": ["createAt", "-createAt", "updateAt", "-updateAt"],
              "type": "string"
            }
          },
          {
            "name": "labels",
            "required": false,
            "in": "query",
            "description": "标签",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "_limit",
            "required": false,
            "in": "query",
            "description": "分页大小",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "_offset",
            "required": false,
            "in": "query",
            "description": "分页偏移",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "name_like",
            "required": false,
            "in": "query",
            "description": "名称 模糊查询",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A paged array of namespace scopes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Namespace"
                  }
                }
              }
            }
          }
        },
        "tags": ["namespace"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Auth API Server",
    "description": "Auth API for auth service",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateNamespaceDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "string",
            "description": "额外数据"
          },
          "desc": {
            "type": "string",
            "description": "描述"
          },
          "labels": {
            "description": "标签",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "name": {
            "type": "string",
            "description": "名称"
          },
          "key": {
            "type": "string",
            "description": "所属命名空间\n\n/ 是分隔符，不允许出现在 namespace 的 ns 中，^[a-zA-Z][a-zA-Z0-9._-]{0,30}$"
          },
          "parent": {
            "type": "string",
            "description": "父级命名空间"
          },
          "registerDefaultRoles": {
            "description": "默认角色",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "passwordRegExp": {
            "type": "string",
            "description": "自定义密码规则"
          }
        },
        "required": ["name", "key"]
      },
      "Namespace": {
        "type": "object",
        "properties": {
          "data": {
            "type": "string",
            "description": "额外数据"
          },
          "desc": {
            "type": "string",
            "description": "描述"
          },
          "labels": {
            "description": "标签",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "name": {
            "type": "string",
            "description": "名称"
          },
          "key": {
            "type": "string",
            "description": "所属命名空间\n\n/ 是分隔符，不允许出现在 namespace 的 ns 中，^[a-zA-Z][a-zA-Z0-9._-]{0,30}$"
          },
          "parent": {
            "type": "string",
            "description": "父级命名空间"
          },
          "isScope": {
            "type": "boolean",
            "description": "是否为 Scope"
          },
          "ns": {
            "type": "string",
            "description": "namesapce 全路径"
          },
          "registerDefaultRoles": {
            "description": "默认角色",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "passwordRegExp": {
            "type": "string",
            "description": "自定义密码规则"
          },
          "id": {
            "type": "string",
            "description": "Entity id"
          },
          "createAt": {
            "format": "date-time",
            "type": "string",
            "description": "Entity created at when"
          },
          "updateAt": {
            "format": "date-time",
            "type": "string",
            "description": "Entity updated at when"
          },
          "createBy": {
            "type": "string",
            "description": "Entity created by who"
          },
          "updateBy": {
            "type": "string",
            "description": "Entity updated by who"
          }
        },
        "required": ["name", "key", "isScope", "ns", "id"]
      },
      "UpdateNamespaceDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "string",
            "description": "额外数据"
          },
          "desc": {
            "type": "string",
            "description": "描述"
          },
          "labels": {
            "description": "标签",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "name": {
            "type": "string",
            "description": "名称"
          },
          "registerDefaultRoles": {
            "description": "默认角色",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "passwordRegExp": {
            "type": "string",
            "description": "自定义密码规则"
          }
        }
      },
...
    }
  }
}